1. Allow publishing a virtual program (compared to a direct program in Yukon) 2. Added classes for tracking hardware addressing information

+++ yukon-client/stars/src/com/cannontech/database/data/stars/customer/CustomerAccount.java
@@ -65,7 +65,7 @@ public class CustomerAccount extends DBPersistent {
 		
 		// Delete program events
         com.cannontech.database.data.stars.event.LMProgramEvent.deleteAllLMProgramEvents(
-				getCustomerAccount().getAccountID() );
+				getCustomerAccount().getAccountID().intValue() );
     	
     	// hardware configuration has already been deleted, so we just need to use the DB object here
 		com.cannontech.database.data.stars.appliance.ApplianceBase app =
